japanese/mh: include errno.h for errno
errno must be accessed by a macro from errno.h.

Due to the sheer number of instances use REINPLACE_CMD to replace most
instances of "extern int errno;" with "#include <errno.h>".  In a few
places the errno decleration was a function-scope global where
including a header didn't make sense patch those out (the files already
included errno.h).  This isn't the most elegant, but it's the least of
the issues with this code.